+++ Sokoban/assets/lastlevel.json
@@ -1,0 +1,1 @@
+[10, [[]], []]
